Fixed PBCs with rotations and vector differences
Positions to be rotated were not being unwrapped relative to the
rotation center. The sign was wrong for taking the difference of two
vectors. I also made some arguments that are vectors into references to
try and make it more efficient (although profiling shows it made little
differene). I should probably insert const markers on these as well.