Fixed bugs in calculations relatd to monomers
The pairwise interactions (calculating energy and simple check for
interaction) was being done wrong; was missing some of the off-diagonal
particle-particle pairs. The calcluation of monomer centers forgot to
take unwrap them with respect to each other. Also added (particle
version was in previous commit) a method to put current configuration
to trial.